Move python code up (to be re-used by nidm-experiment component)
