+++ scripts/OwlReader.py
